move code in FileOps that relies on NetOps to NetOps, put default resource in Visualizer when user has no file selected
