+++ src/components/viewport-workers/NetOps.js
@@ -1,3 +1,20 @@
+import io from 'socket.io-client'
+
+let boneNames = {
+    LUA: "upperarm_l", 
+    LLA: "lowerarm_l", 
+    RUA: "upperarm_r", 
+    RLA: "lowerarm_r", 
+    BACK: "spine_02", /** IMPORTANT */
+    LSHOE: "foot_l", 
+    RSHOE: "foot_r",
+    ROOT: "_rootJoint",
+    RUL: "right_upper_leg",
+    LUL: "left_upper_leg",
+    RLL: "right_lower_leg",
+    LLL: "left_lower_leg"
+}
+
 // PARENT XHR METHOD
 
 async function doXHR(method, url) {
@@ -105,6 +122,141 @@ export function downloadFile(props, fname) {
     }
 
     x.send();
+}
+
+
+//TODO: Should move this to NetOps and merge with existing concept for file downloads
+function getFile(fileName) {
+    let dataPromise = getFilePart(fileName, 'data');
+    let metadataPromise = getFilePart(fileName, 'metadata');
+    return Promise.all([dataPromise, metadataPromise]);
+}
+
+
+function getFilePart(fileName, type) {
+    return new Promise((resolve, reject) => {
+        if (type !== 'data' && type !== 'metadata') {
+            console.log("Inappropriate data type: Should either be data or metadata.");
+        }   
+        let dataReq = new XMLHttpRequest();
+        dataReq.onload = (event => {
+            console.log(dataReq);
+            switch(dataReq.status) {
+                case (200):
+                    resolve(dataReq.responseText);
+                    break;
+                case (400):
+                    reject("Invalid data type request!");
+                    break;
+                case (404):
+                    reject("The target file was not found.");
+                    break;
+            }
+        });
+        dataReq.onerror = (() => {
+            reject();
+        })
+        const targetURL = ("/api/uploadedfiles?")
+        const params = new URLSearchParams();
+        params.set('file', fileName);
+        params.set('type', type);
+        params.set('accessCode', 'password_wrong');
+        console.log(params.toString());
+        const target = targetURL + params.toString();
+        dataReq.open("GET", target);
+        // TODO: Right now the response is in the default 'text' format, but it might
+        // be more appropriate to use another format.
+        dataReq.send();
+        console.log("GET request has been sent: ");
+    });
+}
+
+export function subscribeToFile(props, mySelectedFile) {
+    const URL = "http://localhost:3000"; // This needs to be flexible based on whether we're locally running
+    const socket = io(URL, { 
+        autoConnect: false,
+        auth: {
+            username: "placeholder_username"
+        },
+        query: {
+            "file": mySelectedFile
+        },
+    });
+
+    socket.on("Processing data", () => {
+        console.log("Received processing data message");
+        props.setFileStatus({ status: "Processing data" }); // What should I use as a stand-in for changing status here? Probably props.setStatus("Processing Data"), or just use a flag
+    });
+
+    socket.on("File ready", () => {
+        props.setFileStatus({ status: "Loading file" });
+        console.log(mySelectedFile);
+        getFileList(props);
+        getFile(mySelectedFile)
+        .then((responses) => {
+            console.log(responses[0]); // Data file
+            console.log(responses[1]); // Metadata file
+
+            //When we get the files, we should use the original code to assign them as quaternions (after decoding from metadata)
+            // Lifted from NetOps.js temporarily
+            let inputArray = responses[0].split("\n");
+            let linesArray = [];
+    
+            for (let i = 2; i < inputArray.length - 1; i++) {
+                linesArray.push(inputArray[i].split(" ").filter(item => item !== '\r'));
+            }
+    
+            props.data.current = linesArray;
+            props.setDownloading(false);
+            props.outgoingRequest = false;
+
+            props.fileMetadata.current = JSON.parse(responses[1]);
+
+            socket.disconnect();
+            props.setFileStatus({ status: "Loading models" });
+            props.initializeScene().then((newSceneInfo)=> { // This call gets us way too nested. This should be extracted as a function.
+                props.setSceneInfo({
+                    scene: newSceneInfo.scene,
+                    model: newSceneInfo.model,
+                    camera: null,
+                    renderer: newSceneInfo.renderer,
+                });                
+                
+                let modelBoneList = Object.getOwnPropertyNames(boneNames);
+
+                let bones = [];
+                for (let i = 0; i < modelBoneList.length; i++) {
+                    bones[modelBoneList[i]] = newSceneInfo.model.getObjectByName(boneNames[modelBoneList[i]])
+                }
+                props.onLoadBones(bones)
+                // props.batchUpdate("RUA", [0,0,0,1]);
+                props.setTimeSliderValue(0);
+                props.setFileStatus({ status: "Complete" }); // Determining the next stage by completing the previous stage forces sequential loading. Try using progress flags.
+            });
+        })
+        .catch((error) => {
+            // this.setState({
+            //     errorMessage: "Encountered an error retrieving the file from the server. Try reloading or resubmitting."
+            // });
+            props.setFileStatus({ status: "Error", message: "The file could not be retrieved from the server. Try refreshing or resubmitting." });
+            socket.disconnect();
+        });
+    });
+
+    socket.on("File missing", () => {
+        console.log("File doesn't exist.");
+        props.setFileStatus({ status: "Error", message: "The requested file doesn't exist. Try reselecting the file or resubmitting." });
+        socket.disconnect();
+    })
+
+    socket.onAny((event, ...args) => {
+        console.log("Received data through socket.");
+        console.log(event, args);
+    });
+
+    socket.connect();
+    console.log("Connected with socket.");
+    props.setFileStatus({ status: "Contacting server" });
 }
 
 export function downloadMetafile(props, selectedFilename) {
